Move EF trigger to its own file
